added linked_list.h header file, moved tests to a separate files
